Added support for dumping binary data from ucode; changed ucode version code.
You can use the --dump-ucode and --dump-path options to configure
dumping the RAM section from the ucode.

Also changed is the way we get the version information out of the
buffer. We also take care of endianess now. This may, however, break the
recent hacks I've been using to take care of the microcode versions.