Added an 'install' target and config.h.
It should now be possible to build and install the firmware loader onto
the local system. You can override the installation path when running
make install.

There's still a bit of an issue with creating Debian packages; for some
reason, dh_make runs 'make install' but ends up forcing it to install in
/bin rather than /usr/bin.

Now that I think about it, we probably should be using /usr/sbin instead.