Path fixes within Makefile.
Thanks to Willem van Engen for picking this up. Basically, we now make
sure that we don't override DESTDIR, and make sure that we read out both
it and PREFIX when generating an installation path.

This should also fix building with dpkg.

It's now also possible to easily override the installation name of the
loader executable (currently r5u87x-loader) and the location of the
installed firmware.