Add make rules target, rename UCODEDIR to UCODE_PATH, general cleanup of Makefile
Added the ability to generate a udev rules file based off the template
located in the contrib/ directory. If built, this will also be installed
when running the 'install' target.

The loader tool is now installed into the PREFIX/sbin directory, rather
than PREFIX/bin.

Removed the hard coded list of firmware files; generate them from
listing the ucode directory and using xargs to separate them with spaces.

Also, pass across the microcode path to the compiler from the Makefile
rather than keep one copy in the Makefile and another in config.h.